Display a toast (Toastify) when the user enters an invalid number
The reason there is no invalid toast displaying when the user enters in a letter is because the type=number. type=number is used so that when a user wants to enter input on mobile, a numeric input appears.

This input type doesn't allow letters to be typed, so the event doesn't get fired, which means the popup does not show.

There is a workaround in the form of the inputmode attribute, but it is not supported by all browsers.

I chose to keep it the way it is because I hope users will be smart enough to know that only numbers can be inputted into a sudoku board.