updated code to make findSolution asynchronous
I originally thought that async meant nonblocking, which I just found out is not true. I wanted my code to be nonblocking so the user could continue to use the website without the screen freezing. Unfortunately I will have to settle for what I have or find a different solution.