Updated how the board gets disabled after solution is found
I dont want the user to be able to clear the board after a solution is found, in order to do this I will add a reset button that they will have to press. Originally I handled this when the cells values are being updated. This caused a bug where only the blue letters would be disabled, but after switching to this new method, all cells are updated.